feat: restructure nav — AccountActivity as main launcher with drawer sidebar and account switcher
Make collections (AccountActivity) the main entry point. The account list
moves into a drawer sidebar with a Gmail-style account switcher in the nav
header. AccountsActivity becomes a simple redirect safety net.

New ActiveAccountManager utility persists the active account in SharedPreferences,
with fallback to the first available account when no preference is saved.

Bug fixes included:
- Fix sidebar icons invisible in light theme (ic_add_light white fill → black for Material tinting)
- Fix nav header cut off (fitsSystemWindows + proper padding instead of hardcoded 32dp)
- Fix login form contrast (add textColorPrimary to TextInputEditText fields)
- Fix "Status unavailable" subscription card (hide card entirely when status unknown)
- Fix crash on launch when EXTRA_ACCOUNT missing (ActiveAccountManager fallback)
