+++ android/app/src/main/res/layout/login_credentials_fragment.xml
@@ -42,6 +42,7 @@
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:hint="@string/login_email_address"
+                    android:textColor="?android:attr/textColorPrimary"
                     android:autofillHints="emailAddress"
                     android:inputType="textEmailAddress"/>
             </com.google.android.material.textfield.TextInputLayout>
@@ -59,6 +60,7 @@
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:fontFamily="monospace"
+                    android:textColor="?android:attr/textColorPrimary"
                     android:autofillHints="password"
                     android:inputType="textPassword"
                     android:hint="@string/login_password"/>
