+++ android/app/src/main/res/layout/nav_header_accounts.xml
@@ -2,42 +2,115 @@
 <!-- Modified by Silent Suite -->
 
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-              xmlns:tools="http://schemas.android.com/tools"
-              android:layout_width="match_parent"
-              android:layout_height="wrap_content"
-              android:background="@drawable/nav_header_bg"
-              android:gravity="bottom"
-              android:orientation="vertical"
-              android:paddingStart="16dp"
-              android:paddingEnd="16dp"
-              android:paddingTop="32dp"
-              android:paddingBottom="16dp">
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:background="@drawable/nav_header_bg"
+    android:orientation="vertical"
+    android:fitsSystemWindows="true">
 
-    <ImageView
-        android:layout_width="56dp"
-        android:layout_height="56dp"
-        android:layout_marginBottom="8dp"
-        android:src="@mipmap/ic_launcher"
-        tools:ignore="ContentDescription"/>
-
-    <TextView
-        android:id="@+id/nav_user_email"
+    <!-- Header section with logo, email, dropdown arrow -->
+    <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginTop="8dp"
-        android:text="@string/app_name"
-        android:textColor="@color/nav_header_email"
-        android:textSize="16sp"
-        android:textStyle="bold"
-        android:ellipsize="end"
-        android:maxLines="1"/>
+        android:orientation="vertical"
+        android:gravity="bottom"
+        android:paddingStart="16dp"
+        android:paddingEnd="16dp"
+        android:paddingTop="16dp"
+        android:paddingBottom="16dp">
 
-    <TextView
-        android:layout_width="wrap_content"
+        <ImageView
+            android:layout_width="56dp"
+            android:layout_height="56dp"
+            android:layout_marginBottom="8dp"
+            android:src="@mipmap/ic_launcher"
+            tools:ignore="ContentDescription"/>
+
+        <LinearLayout
+            android:id="@+id/nav_account_header"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:orientation="horizontal"
+            android:gravity="center_vertical"
+            android:layout_marginTop="8dp">
+
+            <TextView
+                android:id="@+id/nav_user_email"
+                android:layout_width="0dp"
+                android:layout_height="wrap_content"
+                android:layout_weight="1"
+                android:text="@string/app_name"
+                android:textColor="@color/nav_header_email"
+                android:textSize="16sp"
+                android:textStyle="bold"
+                android:ellipsize="end"
+                android:maxLines="1"/>
+
+            <ImageView
+                android:id="@+id/nav_account_dropdown"
+                android:layout_width="24dp"
+                android:layout_height="24dp"
+                android:layout_marginStart="8dp"
+                android:src="@android:drawable/arrow_down_float"
+                android:visibility="gone"
+                tools:ignore="ContentDescription"/>
+
+        </LinearLayout>
+
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="2dp"
+            android:textColor="@color/nav_header_subtitle"
+            android:textSize="12sp"
+            android:text="@string/navigation_drawer_subtitle"/>
+
+    </LinearLayout>
+
+    <!-- Expandable account list section -->
+    <LinearLayout
+        android:id="@+id/nav_account_list"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginTop="2dp"
-        android:textColor="@color/nav_header_subtitle"
-        android:textSize="12sp"
-        android:text="@string/navigation_drawer_subtitle"/>
+        android:orientation="vertical"
+        android:background="?android:attr/colorBackground"
+        android:visibility="gone">
+
+        <!-- Account rows will be added programmatically here -->
+
+        <View
+            android:layout_width="match_parent"
+            android:layout_height="1dp"
+            android:background="?android:attr/listDivider"/>
+
+        <LinearLayout
+            android:id="@+id/nav_add_account_row"
+            android:layout_width="match_parent"
+            android:layout_height="48dp"
+            android:orientation="horizontal"
+            android:gravity="center_vertical"
+            android:paddingStart="16dp"
+            android:paddingEnd="16dp"
+            android:background="?android:attr/selectableItemBackground">
+
+            <ImageView
+                android:layout_width="24dp"
+                android:layout_height="24dp"
+                android:src="@drawable/ic_add_light"
+                android:tint="?android:attr/textColorPrimary"
+                tools:ignore="ContentDescription,UseAppTint"/>
+
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginStart="16dp"
+                android:text="@string/navigation_drawer_add_account"
+                android:textColor="?android:attr/textColorPrimary"
+                android:textSize="14sp"/>
+
+        </LinearLayout>
+
+    </LinearLayout>
 
 </LinearLayout>
